refactor: address feedback about IN_BROWSER
